forum/backend/pyproject.toml
[project]
name = "forum-backend"
version = "0.1.0"
description = "FastAPI Backend for Forum"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "python-multipart",
    "python-dotenv",
]

[tool.uv]
managed = true
package = false
